feat: AutomationDebug + autopatch demo module
Demo-specific automation code for the mini-rig funder demo.
Includes autopatch loop, cell detection, nucleus extraction demo,
and related utilities.
