+++ acq4/modules/AutomationDebug/window.ui
@@ -77,6 +77,19 @@
           </property>
          </widget>
         </item>
+        <item>
+         <widget class="QPushButton" name="showRoisBtn">
+          <property name="text">
+           <string>Show Cell ROIs</string>
+          </property>
+          <property name="checkable">
+           <bool>true</bool>
+          </property>
+          <property name="checked">
+           <bool>true</bool>
+          </property>
+         </widget>
+        </item>
        </layout>
       </item>
       <item row="1" column="0" colspan="4">
